Moving Autonne and Takagi from symplectic to decompositions
